requirements-dev.txt
pytest
black
djlint
GitPython
